+++ lib/src/expressions/match_string_expression.dart
@@ -1,0 +1,19 @@
+import 'expression.dart';
+
+class MatchStringExpression extends Expression {
+  final String string;
+
+  MatchStringExpression({
+    required this.string,
+  });
+
+  @override
+  T accept<T>(ExpressionVisitor<T> visitor) {
+    return visitor.visitMatchString(this);
+  }
+
+  @override
+  String toString() {
+    return '@matchString()';
+  }
+}